% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parameters.R
\name{BoolParameter}
\alias{BoolParameter}
\title{BoolParameter}
\usage{
BoolParameter(name, id = NULL, default = NULL, optional = FALSE,
  include_in_default = TRUE)
}
\arguments{
\item{name}{Human-readable name for the parameter, must match argument of scoring function}

\item{id}{(optional) Unique id for parameter (no spaces allowed)}

\item{default}{(optional) Initial value for parameter. If not defined, it will be FALSE.}

\item{optional}{(optional) Whether the parameter can be omitted in a Configuration}

\item{include_in_default}{(optional) Whether an optional parameter will be included in the default Configuration}
}
\description{
A parameter which will be assigned either TRUE or FALSE.
}
