% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parameters.R
\name{ConstantParameter}
\alias{ConstantParameter}
\title{ConstantParameter}
\usage{
ConstantParameter(name, value, id = NULL, optional = FALSE,
  include_in_default = TRUE)
}
\arguments{
\item{name}{Human-readable name for the parameter, must match argument of scoring function}

\item{value}{Can be string, numeric or boolean}

\item{id}{(optional) Unique id for parameter (no spaces allowed)}

\item{optional}{(optional) Whether the parameter can be omitted in a Configuration}

\item{include_in_default}{(optional) Whether an optional parameter will be included in the default Configuration}
}
\description{
A parameter which will always be assigned a specified value.
}
