% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{FloatParameter}
\alias{FloatParameter}
\title{FloatParameter}
\usage{
FloatParameter(name, minimum, maximum, distribution = "Uniform",
  id = NULL, default = NULL, optional = FALSE,
  include_in_default = TRUE, cyclical = FALSE)
}
\arguments{
\item{name}{Human-readable name for the parameter, must match argument of scoring function}

\item{minimum}{Smallest allowed value}

\item{maximum}{Largest allowed value}

\item{distribution}{(optional) Distribution of parameter values ("Uniform" or "LogUniform").}

\item{id}{(optional) Unique id for parameter (no spaces allowed)}

\item{default}{(optional) Initial value for parameter (must be within the allowed range). If not defined, it will be the midpoint of the range.}

\item{optional}{(optional) Whether the parameter can be omitted in a Configuration}

\item{include_in_default}{(optional) Whether an optional parameter will be included in the default Configuration}

\item{cyclical}{(optional) If TRUE, OPTaaS will select values from a period starting at the `minimum` (inclusive) and ending at the `maximum` (exclusive). Values near the minimum and maximum will be considered to be close, as if they were on a circle.

Note: if `cyclical` is TRUE, `distribution` will be ignored. Also, if any of your parameters are cyclical, all your parameters must be Floats, Constants or Groups (other types are not currently supported), and none of them can be `optional`.}
}
\description{
A parameter which can be assigned float values within a specified range (inclusive).
}
