% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{GroupParameter}
\alias{GroupParameter}
\title{GroupParameter}
\usage{
GroupParameter(name, items, id = NULL, default = NULL,
  optional = FALSE, include_in_default = TRUE)
}
\arguments{
\item{name}{Human-readable name for the parameter, must match argument of scoring function}

\item{items}{List of parameters to be included in group (can be empty).}

\item{id}{(optional) Unique id for parameter (no spaces allowed)}

\item{default}{(optional) Initial value for parameter (must be present in the choices list). If not defined, it will be the first item from the list.}

\item{optional}{(optional) Whether the parameter can be omitted in a Configuration}

\item{include_in_default}{(optional) Whether an optional parameter will be included in the default Configuration}
}
\description{
A set of parameters, all of which will be included in each Configuration (unless any of them are optional).
}
