% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parameters.R
\name{IntParameter}
\alias{IntParameter}
\title{IntParameter}
\usage{
IntParameter(name, minimum, maximum, distribution = "Uniform",
  id = NULL, default = NULL, optional = FALSE,
  include_in_default = TRUE)
}
\arguments{
\item{name}{Human-readable name for the parameter, must match argument of scoring function}

\item{minimum}{Smallest allowed value}

\item{maximum}{Largest allowed value}

\item{distribution}{(optional) Distribution of parameter values ("Uniform" or "LogUniform").}

\item{id}{(optional) Unique id for parameter (no spaces allowed)}

\item{default}{(optional) Initial value for parameter (must be within the allowed range). If not defined, it will be the midpoint of the range.}

\item{optional}{(optional) Whether the parameter can be omitted in a Configuration}

\item{include_in_default}{(optional) Whether an optional parameter will be included in the default Configuration}
}
\description{
A parameter which can be assigned integer values within a specified range (inclusive).
}
