% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Objective.R
\name{Objective}
\alias{Objective}
\title{Objective}
\usage{
Objective(id, goal = NULL, min_known_score = NULL,
  max_known_score = NULL)
}
\arguments{
\item{id}{Unique id for the objective}

\item{goal}{(optional, defaults to "max") Use "min" if the goal is to minimise the score for this objective.}

\item{min_known_score}{(optional) The lowest possible score value, if known.}

\item{max_known_score}{(optional) The highest possible score value, if known.}
}
\description{
Specification of an objective for a multi-objective Task.
}
