% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPTaasClient.R
\docType{data}
\name{OPTaaSClient}
\alias{OPTaaSClient}
\title{OPTaaSClient}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
OPTaaSClient
}
\description{
Sets up a connection to OPTaaS and allows you to create a Task, retrieve existing Tasks etc.
}
\examples{
\dontrun{
client <- OPTaaSClient$new("Your OPTaaS URL", "Your OPTaaS API Key")

# Create a task:
task <- client$create_task(
    title="Dummy task",
    parameters=list(
        list(type="boolean", name="dummy_bool")
    ),
    # optional arguments
    goal="min",  # default is "max"
    target_score=100,  # optimal score (if known)
    initial_configurations=5,  # default is 10
    random_seed=123,  # use only if you need reproducible results
    user_defined_data=list(any=data)  # any other data you wish to store
)
}
}
\keyword{datasets}
