% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPTaaSSession.R
\docType{data}
\name{OPTaaSSession}
\alias{OPTaaSSession}
\title{OPTaaSSession}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
OPTaaSSession
}
\description{
Makes requests to OPTaaS and returns responses.
}
\keyword{datasets}
