% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Result.R
\docType{data}
\name{Result}
\alias{Result}
\title{Result}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Result
}
\arguments{
\item{configuration}{The Configuration used to obtain this Result.}

\item{score}{The score obtained. 
For multi-objective tasks, a named list of scores for each objective (using the objective IDs as keys).}

\item{variance}{(optional, defaults to 0) Variance associated with the score. 
For multi-objective tasks, a named list  of variances for each objective (using the objective IDs as keys).}

\item{error}{Any data related to an error encountered while calculating the score.}

\item{user_defined_data}{(optional) Any other data you wish to store as part of this Result.}
}
\description{
The result obtained by taking a Configuration and running it through your scoring function.
}
\details{
Must contain either a score value or an error.
You cannot specify variance without score.
}
\keyword{datasets}
