% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApiKey.R
\docType{data}
\name{ApiKey}
\alias{ApiKey}
\title{ApiKey}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ApiKey
}
\description{
An API Key that can be used to interact with OPTaaS.
}
\section{Fields}{

\describe{
\item{\code{id}}{The actual API Key.}

\item{\code{role}}{The role assigned to this key.}

\item{\code{expired}}{Whether this key has expired, i.e. can no longer be used.}

\item{\code{json}}{The full JSON representation of this key in OPTaaS.}
}}

\keyword{datasets}
