% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPTaasClient.R
\docType{data}
\name{OPTaaSClient}
\alias{OPTaaSClient}
\title{OPTaaSClient}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
OPTaaSClient
}
\description{
Sets up a connection to OPTaaS and allows you to create a Task, retrieve existing Tasks etc.
}
\examples{
\dontrun{

client <- OPTaaSClient$new("Your OPTaaS URL", "Your OPTaaS API Key")

# Create a task:
task <- client$create_task(
    title="Dummy task",
    parameters=list(
        BoolParameter("my_bool"),
        CategoricalParameter("my_cat", values=list("a", "b", "c"), id="cat"),
        ChoiceParameter('ints_or_floats', choices=list(
            GroupParameter('ints', items=list(
                IntParameter('my_int', minimum=0, maximum=20, id="int"),
                IntParameter('optional_int', minimum=-10, maximum=10, optional=TRUE, id="opt_int")
            )),
            GroupParameter('floats', items=list(
                FloatParameter('float1', minimum=0, maximum=1),
                FloatParameter('float2', minimum=0.5, maximum=4.5)
            ))
        ))
    ),
    
    # for single-objective tasks
    goal="min",  # optional, default is "max"
    min_known_score=0, max_known_score=100,  # optional
    
    # for multi-objective tasks
    objectives=list(
        Objective(id="objective1", goal="min", min_known_score=0, max_known_score=100)
        Objective(id="objective2"),  # defaults to goal="max"
    ),
    
    # optional arguments
    initial_configurations=5,  # default is 10
    random_seed=123,  # use only if you need reproducible results
    user_defined_data=list(any="data"),  # any other data you wish to store
)

# Get a task:
task <- client$get_task("Task ID")

# Get all tasks:
tasks <- client$get_all_tasks()
}
}
\keyword{datasets}
