% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Task.R
\docType{data}
\name{Task}
\alias{Task}
\title{Task}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Task
}
\description{
Allows you to access Task attributes and perform all Task-related functions.
}
\examples{
\dontrun{
task.id  # The unique ID of the Task
task.json  # The full JSON representation of the Task in OPTaaS

task.delete()  # Delete the task
}
}
\keyword{datasets}
